% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carto_valencia}
\alias{carto_valencia}
\title{Cartografía}
\format{Un eespatial
\describe{
  \item{id}{id de la sección}
  \item{cusec}{código...}
}}
\usage{
carto_valencia
}
\description{
Cartografía de la ciudad de Valencia
}
\keyword{datasets}
