% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pinta_mapas}
\alias{pinta_mapas}
\title{Pintar mapas}
\usage{
pinta_mapas(cartografia, variable, n_grupos, title = NULL)
}
\arguments{
\item{cartografia}{objeto de la clase Spatial}

\item{variable}{vector num}

\item{n_grupos}{valor numérico}

\item{title}{título del gráfico}
}
\description{
Dado un objeto tatatatatata
}
