% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{privacion_valencia}
\alias{privacion_valencia}
\title{Indicadroes de priv}
\format{Un eespatial
\describe{
  \item{seccion_censal}{seccion censal}
  \item{desempt}{desempleo}
}}
\usage{
privacion_valencia
}
\description{
Indicadores privacion
}
\keyword{datasets}
